backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
daphne==4.0.0
redis==5.0.1
django-cors-headers==4.3.1
APScheduler==3.10.4
psycopg2-binary==2.9.11
akshare>=1.11.0
pandas>=2.0.0